Obsolete ComplexLifecycleTests by ensuring all scenarios are covered by LifecycleTests.
